optimized-multi-coin-tipbot/requirements.txt
# Optimized Multi-Coin Tipbot Requirements
# Only essential dependencies for your 4 coins

# Core Telegram Bot
python-telegram-bot==20.7
aiofiles==24.1.0

# Cryptography and Security
cryptography==41.0.7
bcrypt==4.1.2
pyotp==2.9.0
mnemonic==0.21

# QR Code Generation
qrcode[pil]==8.2
Pillow==10.1.0

# Database
aiosqlite==0.19.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Async and Utilities
schedule==1.2.2
psutil==5.9.6

# Web Dashboard
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3

# Environment and Configuration
python-dotenv==1.0.0

# Time and Date
python-dateutil==2.8.2
pytz==2023.3

# Validation
validators==0.22.0